move extension test into separate file
